requirements.txt
websockets==11.0.3
pyserial-asyncio==0.5
